% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ogreleases_data}
\alias{ogreleases_data}
\title{Selection of Government of Ontario news releases from 2019-2020}
\format{
A data frame with 1503 rows and 7 variables.
}
\source{
\url{https://news.ontario.ca/en/}
}
\usage{
ogreleases_data
}
\description{
Selection of Government of Ontario news releases from 2019-2020
}
\keyword{datasets}
